refactored the folder structure some more
