+++ src/main/java/com/example/demo/outputs/repositories/MediaRepository.java
@@ -1,4 +1,4 @@
-package com.example.demo.reality;
+package com.example.demo.outputs.repositories;
 
 import com.example.demo.entities.Media;
 import org.springframework.data.jpa.repository.JpaRepository;
