+++ src/main/java/com/example/demo/outputs/repositories/RealityRepository.java
@@ -1,6 +1,6 @@
-package com.example.demo.reality;
+package com.example.demo.outputs.repositories;
 
-import com.example.demo.outputs.adapters.entities.RealityEntity;
+import com.example.demo.outputs.entities.RealityEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
