+++ src/main/java/com/example/demo/user/UserMapper.java
@@ -1,10 +1,8 @@
 package com.example.demo.user;
 
 
-import com.example.demo.domain.models.Reality;
 import com.example.demo.domain.models.User;
-import com.example.demo.outputs.adapters.entities.RealityEntity;
-import com.example.demo.outputs.adapters.entities.UserEntity;
+import com.example.demo.outputs.entities.UserEntity;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
