+++ src/main/java/com/example/demo/user/UserRepository.java
@@ -1,6 +1,6 @@
 package com.example.demo.user;
 
-import com.example.demo.outputs.adapters.entities.UserEntity;
+import com.example.demo.outputs.entities.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
